Updated functions and index.js
